@@ -53,12 +53,14 @@
     "react-dom": "^16.13.1",
     "react-feather": "^2.0.8",
     "react-input-range": "^1.3.0",
+    "react-intersection-observer": "^9.4.0",
     "react-intl": "^5.15.8",
     "react-loading-skeleton": "^2.1.1",
     "react-meta-tags": "^1.0.1",
     "react-redux": "^7.2.1",
     "react-responsive-carousel": "^3.2.22",
-    "react-router-dom": "^5.2.0",
+    "react-router": "6",
+    "react-router-dom": "6",
     "react-scripts": "3.4.1",
     "react-slick": "^0.28.1",
     "react-spinners": "^0.11.0",
@@ -70,7 +72,7 @@
     "sass": "^1.49.7",
     "slick-carousel": "^1.8.1",
     "styled-components": "^5.1.1",
-    "typescript": "^3.9.7",
+    "typescript": "^4.7.4",
     "use-wallet": "OGBoogieTomb/use-wallet",
     "web-vitals": "^1.0.1",
     "web3": "^1.2.11",
@@ -110,7 +112,7 @@
   "devDependencies": {
     "@types/numeral": "0.0.28",
     "@types/react-dom": "^16.9.8",
-    "@types/react-router-dom": "^5.1.5",
+    "@types/react-router-dom": "^5.3.3",
     "@types/styled-components": "^5.1.2",
     "css-loader": "^4.3.0",
     "prettier": "2.3.1"
