@@ -1,21 +1,21 @@
 import React from 'react'
-import {useSelector} from "react-redux";
-import {Route, Switch} from 'react-router-dom'
+import { useSelector } from "react-redux";
+import { Route, Routes } from 'react-router-dom'
 
-const Routing = ({routes}) => {
-    const {access} = useSelector(state => state.rootReducer.accessReducer)
+const Routing = ({ routes }) => {
+    const { access } = useSelector(state => state.rootReducer.accessReducer)
 
     return (
-        <Switch>
+        <Routes>
             {
-                routes.map(item =>
+                routes.map((item, i) =>
                     item.role.indexOf(access) !== -1 &&
                     <Route
-                        key={new Date().getTime()} {...item}
+                        key={i} {...item}
                     />
                 )
             }
-        </Switch>
+        </Routes>
     );
 };
 
