@@ -1,16 +1,16 @@
-import React from "react";
-import {routes} from './routing/mainRoutes';
+import React, { Suspense } from "react";
+import { routes } from './routing/mainRoutes';
 
-import {Provider} from 'react-redux'
-import {BrowserRouter} from 'react-router-dom';
+import { Provider } from 'react-redux'
+import { BrowserRouter } from 'react-router-dom';
 // import MetaTags from "react-meta-tags";
 import Routing from "./components/Routing";
-
+import Loader from './components/Loader';
 import styles from './App.module.scss';
 import store from "./state";
-import {UseWalletProvider} from "use-wallet";
+import { UseWalletProvider } from "use-wallet";
 import config from "./config";
-import {RefreshContextProvider} from "./contexts/RefreshContext";
+import { RefreshContextProvider } from "./contexts/RefreshContext";
 import TombFinanceProvider from "./contexts/TombFinanceProvider";
 import ModalsProvider from './contexts/Modals';
 import BanksProvider from './contexts/Banks';
@@ -39,7 +39,7 @@ const App = () => {
             <UseWalletProvider
                 chainId={config.chainId}
                 connectors={{
-                    walletconnect: {rpcUrl: config.defaultProvider},
+                    walletconnect: { rpcUrl: config.defaultProvider },
                     walletlink: {
                         url: config.defaultProvider,
                         appName: 'Sundae Finance',
@@ -48,16 +48,18 @@ const App = () => {
                 }}
             >
                 <Provider store={store}>
-                    <Updaters/>
+                    <Updaters />
                     <RefreshContextProvider>
                         <TombFinanceProvider>
                             <ModalsProvider>
                                 <BanksProvider>
-                                    <BrowserRouter>
-                                        <Page>
-                                            <Routing {...{routes}} />
-                                        </Page>
-                                    </BrowserRouter>
+                                    <Suspense fallback={<Loader />}>
+                                        <BrowserRouter>
+                                            <Page>
+                                                <Routing {...{ routes }} />
+                                            </Page>
+                                        </BrowserRouter>
+                                    </Suspense>
                                 </BanksProvider>
                             </ModalsProvider>
                         </TombFinanceProvider>
